llm-graph/requirements.txt
# Core dependencies
langchain>=0.1.0
langgraph>=0.1.0
networkx>=3.0
numpy>=1.24.0
pandas>=2.0.0

# LLM Integration
openai>=1.0.0
anthropic>=0.8.0

# Local LLM serving
ollama>=0.1.0

# Data processing
pydantic>=2.0.0
pyyaml>=6.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Logging and monitoring
loguru>=0.7.0
wandb>=0.15.0

# Statistical analysis
scipy>=1.10.0
statsmodels>=0.14.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Development
black>=23.0.0
isort>=5.12.0
mypy>=1.4.0
